added debounce to interrupt and changed to falling edge

+++ src/Stepper.cpp
@@ -63,7 +63,7 @@ void Stepper::ccw_step() {
   energize_step(step);
 }
 
-void Stepper::rotate(int32_t rotations) {
+void Stepper::rotate(double rotations) {
   uint32_t num_steps = abs(rotations * steps_per_revolution);
 
   if (rotations < 0) {
